+++ code/crates/network/src/pubsub.rs
@@ -2,18 +2,19 @@ use bytes::Bytes;
 use libp2p::swarm;
 
 use crate::behaviour::Behaviour;
-use crate::{Channel, PubSubProtocol};
+use crate::{Channel, ChannelNames, PubSubProtocol};
 
 pub fn subscribe(
     swarm: &mut swarm::Swarm<Behaviour>,
     protocol: PubSubProtocol,
     channels: &[Channel],
+    channel_names: ChannelNames,
 ) -> Result<(), eyre::Report> {
     match protocol {
         PubSubProtocol::GossipSub => {
             if let Some(gossipsub) = swarm.behaviour_mut().gossipsub.as_mut() {
                 for channel in channels {
-                    gossipsub.subscribe(&channel.to_gossipsub_topic())?;
+                    gossipsub.subscribe(&channel.to_gossipsub_topic(channel_names))?;
                 }
             } else {
                 return Err(eyre::eyre!("GossipSub not enabled"));
@@ -22,7 +23,7 @@ pub fn subscribe(
         PubSubProtocol::Broadcast => {
             if let Some(broadcast) = swarm.behaviour_mut().broadcast.as_mut() {
                 for channel in channels {
-                    broadcast.subscribe(channel.to_broadcast_topic());
+                    broadcast.subscribe(channel.to_broadcast_topic(channel_names));
                 }
             } else {
                 return Err(eyre::eyre!("Broadcast not enabled"));
@@ -37,19 +38,20 @@ pub fn publish(
     swarm: &mut swarm::Swarm<Behaviour>,
     protocol: PubSubProtocol,
     channel: Channel,
+    channel_names: ChannelNames,
     data: Bytes,
 ) -> Result<(), eyre::Report> {
     match protocol {
         PubSubProtocol::GossipSub => {
             if let Some(gossipsub) = swarm.behaviour_mut().gossipsub.as_mut() {
-                gossipsub.publish(channel.to_gossipsub_topic(), data)?;
+                gossipsub.publish(channel.to_gossipsub_topic(channel_names), data)?;
             } else {
                 return Err(eyre::eyre!("GossipSub not enabled"));
             }
         }
         PubSubProtocol::Broadcast => {
             if let Some(broadcast) = swarm.behaviour_mut().broadcast.as_mut() {
-                broadcast.broadcast(&channel.to_broadcast_topic(), data);
+                broadcast.broadcast(&channel.to_broadcast_topic(channel_names), data);
             } else {
                 return Err(eyre::eyre!("Broadcast not enabled"));
             }